Finished adding cronjob for reboot and restart sqlt to tweak.cgi/writetocrnjob.cgi
piCorePlayer.dep file lists the needed packages - is used in do_updatepicoreplayer.cgi.
However we should discus the way we do the update. The process has been changed lately, so now we overwrite the onboot.lst - that means that any other programs/packages the user had installed is now lost. If we exclude the onboot.lst from being overwritten but just add all piCorePlayer need packages as defined in piCorePlayer.dep file - I think it should work for all?

I also added config.cfg as we need to work on the same when we add new variables.