Added resize partition script
Steen, I thought this script might come in handy during development. I don't think piCorePlayer really needs it.

The interesting thing with this script is the way one script behaves in 3 different ways. First the inital screen, then the fdisk screen and then the resize2fs screen. Usualy we would have 2 or 3 scripts to do this.