Greg please have a look - it is not working perfectly - and I can't fix it.
I have problems with the java-script, which was ment to provide an option (like what you did for backup, reboot etc). Here I wanted it to ask if you would use the auto-discovered LMS IP address - press accept - and then it should continue (this part is working).
BUT the script is also executed if you select cancel/or exit in the java-script pop-up window. And this is not what I wanted.

The java-script and the possibility to use auto-discoveres LMS-IP is only available if no LMS-IP address is provided on the Squeezelite page