Added new option to Squeezelite, "set the timeout period for closing of the output when idle" If you want to play with this you should probably use the squeezelite I have compiled which is present in our GIT as well. Triode have added more new options: -U <control>  Unmute ALSA control and set to full volume (not supported with -V) -V <control>  Use ALSA control for volume adjustment, otherwise use software volume adjustment -P <filename> Store the process id (PID) in filename
I haven't implemented these yet as I couldn't see the need for them in piCorePlayer. But if you could see them useful, I will add them - so please tell me.

I removed all the sed command in writetoconfig.cfg as we had defined this as a function in pcp-functions.

Greg. Should we remove these sed-commands from all the writeto?.cgi files as we already have the same sed-commands in pcp-functions?