Trimmed banner.png
Cut and rescaled banner.png
Removed old and test logos.

I have removed width=100% from the banner html so the image displays the actual size of the graphic.

Steen, I noticed my graphics program has doubled the size (kB) of the pCP logo. If you can open and save it back to a smaller size that would be great. If not, no problem.