Inital webgui enable/disable script
I have a feeling that having the Disable GUI button on the Main page is going to cause us a few problems with people just pressing the button to see what it does. We will have to explain how to ssh into pCP to fix it.

Advanced users that want to Disable the GUI will probably be skilled enough to use ssh to both Enable and Disable it. What do you think?