ETHO and WLAN error redirection. Added ANSI colour to boot process.
I think redirecting the WLAN and ETH0 error messages is a better option. It catches all combinations of wifi on/off, wifi adaptor in/out, Model A with no eth0.

I had been working on do_rebootstuff.sh adding the ANSI colouring. It is obviously not finished but if you are also looking at the rebooting process I guess it would be better/easier to test with ANSI colouring in place. What do you think?