Completed xtras_dosfsck.cgi.
This script allows you to check and fix the dirty bit of the boot partition. This should prevent the "scan and repair" messages when the SD card is inserted into PCs. Further testing required.

WARNING: It does download dosfstools.tcz but it has a button to delete this extension after use.

Minor update to xtras_resize.cgi. Changed to mode=99 and added red warning colours.