Revised writetojivelite.cgi
Hi Steen, not fully tested. This is still pretty much the same code you  did, I just rearranged it a little to make it easier to maintain. I have only done 2 test runs and it seems to be ok but needs more testing. I only pushed it because I see you are doing some code updates and knew you would be looking at this soon.