Added logic that automatically sets the correct timezone.
Greg - please have a look at the script I made for detecting a working internet in the do_rebootstuff.
I'm not sure that is the best way, and I see that we have a function in pcp-functions that seems to do a similar test.
But I can't figure out how to use that test for my purpose in do_rebootstuff.

Also I changed the timezone function in pcp-function, so now any changes are avaiable without a reboot. Do you see any problems here?