Fixed pcp_httpd_query_string routine
The $QUERY_STRING was one line and I was using sed/=/="/g
"g" for global, so the "=" in the sysdefault:CARD=ALSA was getting a " stuck in.

So I put a newline between each variable/parameter pairs and changed sed to only change the first occurrence of "=" on each line.