Fixed the stop and force restart of shairport - IN A RUDE WAY...
Somehow the pid number in the /var/run/shairport-sync.pid file was always 2 higher than the running shairport-sync process found using the top command. Therefore, we could not kill the running process by the pid number from the shairport-sync.pid file.
I read it probably have something to do with the shairport-sync was forking out early in the process - whatever that means...
For now I chosed to simply kill the process...