Implemented squeezelite -G and -S options
The squeezelite gpio toggle and power script are working fine. I added a simple powerscript example in /home/tc that can be used for testing purposes. The output will need to be set to the correct /dev.

The squeezelite log file name option is now hardcoded as /tmp/pcp_squeezelite.log. This means it is easier to set debugging on by just using one setting on the squeezelite settings page.