I think this LMSIP works
Steen, if I understand correctly, the old code worked perfectly except if the localhost IP was 127.0.0.1. In this case, we need to use the IP address of the local LAN or WLAN device.

The original code was fine tuned over the last year or two to eliminate all unusual "netstat" combinations that I could find, that's why it looked a little odd, such as apparently repeated greps.