Replacement for the piCore's rotdash.
The rotating dashes don't look good in the web interface. If this script is placed in the $PATH before /usr/bin (ie. /usr/local/sbin) those dashes will be replaced with dots.
-/-|-\ vs .....

As this is waiting for a process to finish I don't see much benefit doing this in C. Do you think it is worth implementing?