I think the sound card selection routine is ready for testing now.
There is one thing that I couldn't figure out - and that was how to show a Sound card with empty spaces in the name.
The sound card name is defined in LISTNAME variable within the pcp-soundcard-functions. But if I used a name with an empty space like "IQaudIO DAC" only IQaudIO was shown in the dropdown list in squeezelite.cgi.
For now I renamed the cards to IQaudIO_DAC