Use custom icon for the piCorePlayer menu item (e588a3524d93310349c8e4e13b389ff5049bfd24).
Unfortunately I haven't found an easy way to define a custom style inside an applet. Therefore this change is looking for a definition of a custom style, using it if available, falling back to default if not. This should allow us to at least add that style to the most important skins without too much effort (the Raspberry Pi display optimized skins).